requirements.txt
Flask==2.3.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.0
Jinja2==3.1.4
requests==2.32.3
python-dotenv==1.0.0
pymysql==1.1.0
flask_cors==4.0.0
pandas
boto3
flask-wtf
cryptography
beautifulsoup4
selenium
twilio
